refactor: extract deep_link as a standalone feature
