+++ src-tauri/src/features/mod.rs
@@ -1,1 +1,2 @@
+pub mod deep_link;
 pub mod thumbnail;
