+++ src-tauri/tauri.conf.json
@@ -67,7 +67,7 @@
 				}
 			],
 			"desktop": {
-				"schemes": ["sendme", "alt-sendme"]
+				"schemes": ["sendme"]
 			}
 		}
 	},
